Version bump, changelog & build

+++ prettier.novaextension/Scripts/main.js
@@ -2,6 +2,8 @@
 
 Object.defineProperty(exports, '__esModule', { value: true });
 
+var main = {};
+
 class ProcessError extends Error {
 	constructor(status, message) {
 		super(message);
@@ -9,7 +11,7 @@ class ProcessError extends Error {
 	}
 }
 
-function showError(id, title, body) {
+function showError$2(id, title, body) {
 	let request = new NotificationRequest(id);
 
 	request.title = nova.localize(title);
@@ -19,7 +21,7 @@ function showError(id, title, body) {
 	nova.notifications.add(request).catch((err) => console.error(err, err.stack));
 }
 
-function showActionableError(id, title, body, actions, callback) {
+function showActionableError$1(id, title, body, actions, callback) {
 	let request = new NotificationRequest(id);
 
 	request.title = nova.localize(title);
@@ -32,14 +34,14 @@ function showActionableError(id, title, body, actions, callback) {
 		.catch((err) => console.error(err, err.stack));
 }
 
-function getConfigWithWorkspaceOverride(name) {
+function getConfigWithWorkspaceOverride$2(name) {
 	const workspaceConfig = getWorkspaceConfig(name);
 	const extensionConfig = nova.config.get(name);
 
 	return workspaceConfig === null ? extensionConfig : workspaceConfig
 }
 
-function observeConfigWithWorkspaceOverride(name, fn) {
+function observeConfigWithWorkspaceOverride$1(name, fn) {
 	let ignored = false;
 	function wrapped(...args) {
 		if (!ignored) {
@@ -66,7 +68,7 @@ function getWorkspaceConfig(name) {
 	}
 }
 
-function handleProcessResult(process, reject, resolve) {
+function handleProcessResult$1(process, reject, resolve) {
 	const errors = [];
 	process.onStderr((err) => {
 		errors.push(err);
@@ -82,7 +84,7 @@ function handleProcessResult(process, reject, resolve) {
 	});
 }
 
-const log = Object.fromEntries(
+const log$3 = Object.fromEntries(
 	['log', 'info', 'warn'].map((fn) => [
 		fn,
 		(...args) => {
@@ -95,18 +97,18 @@ const log = Object.fromEntries(
 );
 
 var helpers = {
-	showError,
-	showActionableError,
-	log,
-	getConfigWithWorkspaceOverride,
-	observeConfigWithWorkspaceOverride,
+	showError: showError$2,
+	showActionableError: showActionableError$1,
+	log: log$3,
+	getConfigWithWorkspaceOverride: getConfigWithWorkspaceOverride$2,
+	observeConfigWithWorkspaceOverride: observeConfigWithWorkspaceOverride$1,
 	ProcessError,
-	handleProcessResult,
+	handleProcessResult: handleProcessResult$1,
 };
 
-const { handleProcessResult: handleProcessResult$1, log: log$1 } = helpers;
+const { handleProcessResult, log: log$2 } = helpers;
 
-async function findPrettier(directory) {
+async function findPrettier$1(directory) {
 	let resolve, reject;
 	const promise = new Promise((_resolve, _reject) => {
 		resolve = _resolve;
@@ -124,7 +126,7 @@ async function findPrettier(directory) {
 		const [path, name, status, extra] = result.trim().split(':');
 		if (!name || !name.startsWith('prettier@')) return resolve(null)
 		if (path === nova.workspace.path) {
-			log$1.info(
+			log$2.info(
 				`You seem to be working on Prettier! The extension doesn't work without Prettier built, so using the built-in Prettier instead.`
 			);
 			return resolve(null)
@@ -136,7 +138,7 @@ async function findPrettier(directory) {
 		});
 	});
 
-	handleProcessResult$1(process, reject, resolve);
+	handleProcessResult(process, reject, resolve);
 	process.start();
 
 	return promise
@@ -154,7 +156,7 @@ async function installPrettier(directory) {
 		cwd: directory,
 	});
 
-	handleProcessResult$1(process, reject, resolve);
+	handleProcessResult(process, reject, resolve);
 	process.start();
 
 	return promise
@@ -164,33 +166,33 @@ var prettierInstallation = async function () {
 	// Try finding in the workspace
 	try {
 		if (nova.workspace.path) {
-			const resolved = await findPrettier(nova.workspace.path);
+			const resolved = await findPrettier$1(nova.workspace.path);
 			if (resolved) {
-				log$1.info(`Loading project prettier at ${resolved.path}`);
+				log$2.info(`Loading project prettier at ${resolved.path}`);
 				return resolved.path
 			}
 		}
 	} catch (err) {
 		if (err.status === 127) throw err
-		log$1.warn('Error trying to find workspace Prettier', err);
+		log$2.warn('Error trying to find workspace Prettier', err);
 	}
 
 	// Install / update bundled version
 	try {
 		const path = nova.path.join(nova.extension.path, 'node_modules', 'prettier');
 
-		const resolved = await findPrettier(nova.extension.path);
+		const resolved = await findPrettier$1(nova.extension.path);
 
 		if (!resolved || !resolved.correctVersion) {
-			log$1.info(`Installing / updating bundled Prettier at ${path}`);
+			log$2.info(`Installing / updating bundled Prettier at ${path}`);
 			await installPrettier(nova.extension.path);
 		}
 
-		log$1.info(`Loading bundled prettier at ${path}`);
+		log$2.info(`Loading bundled prettier at ${path}`);
 		return path
 	} catch (err) {
 		if (err.status === 127) throw err
-		log$1.warn('Error trying to find or install bundled Prettier', err);
+		log$2.warn('Error trying to find or install bundled Prettier', err);
 	}
 };
 
@@ -218,7 +220,6 @@ var prettierInstallation = async function () {
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 
 /**
  * The data structure representing a diff is an array of tuples:
@@ -949,22 +950,23 @@ function find_cursor_edit_diff(oldText, newText, cursor_pos) {
   return null;
 }
 
-function diff(text1, text2, cursor_pos) {
+function diff$1(text1, text2, cursor_pos) {
   // only pass fix_unicode=true at the top level, not when diff_main is
   // recursively invoked
   return diff_main(text1, text2, cursor_pos, true);
 }
 
-diff.INSERT = DIFF_INSERT;
-diff.DELETE = DIFF_DELETE;
-diff.EQUAL = DIFF_EQUAL;
+diff$1.INSERT = DIFF_INSERT;
+diff$1.DELETE = DIFF_DELETE;
+diff$1.EQUAL = DIFF_EQUAL;
 
-var diff_1 = diff;
+var diff_1 = diff$1;
 
+const diff = diff_1;
 const {
 	showError: showError$1,
-	showActionableError: showActionableError$1,
-	log: log$2,
+	showActionableError,
+	log: log$1,
 	getConfigWithWorkspaceOverride: getConfigWithWorkspaceOverride$1,
 } = helpers;
 
@@ -997,12 +999,11 @@ const PRETTIER_OPTIONS = [
 	'vueIndentScriptAndStyle',
 ];
 
-class Formatter {
+class Formatter$1 {
 	constructor() {
 		this.prettierServiceDidExit = this.prettierServiceDidExit.bind(this);
-		this.prettierServiceStartDidFail = this.prettierServiceStartDidFail.bind(
-			this
-		);
+		this.prettierServiceStartDidFail =
+			this.prettierServiceStartDidFail.bind(this);
 
 		this.emitter = new Emitter();
 
@@ -1029,11 +1030,15 @@ class Formatter {
 
 	async start(modulePath) {
 		if (modulePath) this.modulePath = modulePath;
-		if (this.prettierService) return
-
-		log$2.info('Starting Prettier service');
 
 		if (!this._isReadyPromise) this.setupIsReadyPromise();
+		// If we're currently stopping we'll wait for that to complete before starting
+		if (this._isStoppedPromise) {
+			await _isStoppedPromise;
+		}
+
+		if (this.prettierService) return
+		log$1.info('Starting Prettier service');
 
 		this.prettierService = new Process('/usr/bin/env', {
 			args: [
@@ -1061,15 +1066,23 @@ class Formatter {
 
 	stop() {
 		nova.notifications.cancel('prettier-not-running');
-		if (!this._isReadyPromise) return
-		if (!this.prettierService) return
+		if (!this._isReadyPromise || !this.prettierService) return
+		if (this._isStoppedPromise) return
 
-		log$2.info('Stopping Prettier service');
+		log$1.info('Stopping Prettier service');
 
-		this.prettierService.terminate();
+		this._isStoppedPromise = new Promise((resolve) => {
+			this._resolveIsStoppedPromise = resolve;
+		});
+
+		// Stop processing things
 		if (this._resolveIsReadyPromise) this._resolveIsReadyPromise(false);
 		this._isReadyPromise = null;
+		// Actually terminate
+		this.prettierService.terminate();
 		this.prettierService = null;
+
+		return this._isStoppedPromise
 	}
 
 	setupIsReadyPromise() {
@@ -1079,6 +1092,10 @@ class Formatter {
 	}
 
 	prettierServiceDidExit(exitCode) {
+		if (this._resolveIsStoppedPromise) {
+			this._resolveIsStoppedPromise();
+			this._isStoppedPromise = null;
+		}
 		if (!this.prettierService) return
 
 		console.error(`Prettier service exited with code ${exitCode}`);
@@ -1100,7 +1117,7 @@ class Formatter {
 	prettierServiceStartDidFail({ parameters: error }) {
 		this._resolveIsReadyPromise(false);
 
-		showActionableError$1(
+		showActionableError(
 			'prettier-not-running',
 			`Couldn't load Prettier`,
 			`Is your Node up to date? Or did you set an invalid 'Prettier module' path in the extension or project settings? A detailed log of the error is available in the extension console.`,
@@ -1121,7 +1138,7 @@ class Formatter {
 	}
 
 	showServiceNotRunningError() {
-		showActionableError$1(
+		showActionableError(
 			'prettier-not-running',
 			'Prettier stopped running',
 			`Please report report an issue though Extension Library if this problem persits.`,
@@ -1151,7 +1168,7 @@ class Formatter {
 
 		if (shouldApplyDefaultConfig === null) return []
 
-		log$2.info(`Formatting ${document.path}`);
+		log$1.info(`Formatting ${document.path}`);
 
 		const documentRange = new Range(0, document.length);
 		const original = editor.getTextInRange(documentRange);
@@ -1182,7 +1199,7 @@ class Formatter {
 		}
 
 		if (ignored) {
-			log$2.info(`Prettier is configured to ignore ${document.path}`);
+			log$1.info(`Prettier is configured to ignore ${document.path}`);
 			return []
 		}
 
@@ -1194,12 +1211,12 @@ class Formatter {
 					`Prettier doesn't include a Parser for this file and no plugin is installed that does.`
 				);
 			}
-			log$2.info(`No parser for ${document.path}`);
+			log$1.info(`No parser for ${document.path}`);
 			return []
 		}
 
 		if (formatted === original) {
-			log$2.info(`No changes for ${document.path}`);
+			log$1.info(`No changes for ${document.path}`);
 			return []
 		}
 
@@ -1214,7 +1231,7 @@ class Formatter {
 				`prettier.format-on-save.ignored-syntaxes.${document.syntax}`
 			) === true
 		) {
-			log$2.info(
+			log$1.info(
 				`Not formatting (${document.syntax} syntax ignored) ${document.path}`
 			);
 			return null
@@ -1238,6 +1255,7 @@ class Formatter {
 
 			if (
 				!hasConfig &&
+				saving &&
 				getConfigWithWorkspaceOverride$1(
 					'prettier.format-on-save.ignore-without-config'
 				)
@@ -1267,7 +1285,7 @@ class Formatter {
 	}
 
 	async applyResult(editor, original, formatted) {
-		log$2.info(`Applying formatted changes to ${editor.document.path}`);
+		log$1.info(`Applying formatted changes to ${editor.document.path}`);
 
 		const [cursor, edits] = this.diff(
 			original,
@@ -1278,7 +1296,7 @@ class Formatter {
 		if (
 			original !== editor.getTextInRange(new Range(0, editor.document.length))
 		) {
-			log$2.info(`Document ${editor.document.path} was changed while formatting`);
+			log$1.info(`Document ${editor.document.path} was changed while formatting`);
 			return
 		}
 
@@ -1312,7 +1330,7 @@ class Formatter {
 		originalWithCursors += original.slice(lastEnd);
 
 		// Diff
-		return [cursor, diff_1(originalWithCursors, formatted)]
+		return [cursor, diff(originalWithCursors, formatted)]
 	}
 
 	async applyDiff(editor, cursor, edits) {
@@ -1322,10 +1340,10 @@ class Formatter {
 			let toRemove = 0;
 
 			// Add an extra empty edit so any trailing delete is actually run.
-			edits.push([diff_1.EQUAL, '']);
+			edits.push([diff.EQUAL, '']);
 
 			for (const [edit, str] of edits) {
-				if (edit === diff_1.DELETE) {
+				if (edit === diff.DELETE) {
 					toRemove += str.length;
 
 					// Check if the cursors are in here
@@ -1346,9 +1364,9 @@ class Formatter {
 					continue
 				}
 
-				if (edit === diff_1.EQUAL && toRemove) {
+				if (edit === diff.EQUAL && toRemove) {
 					e.replace(new Range(offset, offset + toRemove), '');
-				} else if (edit === diff_1.INSERT) {
+				} else if (edit === diff.INSERT) {
 					e.replace(new Range(offset, offset + toRemove), str);
 				}
 
@@ -1407,46 +1425,47 @@ class Formatter {
 }
 
 var formatter = {
-	Formatter,
+	Formatter: Formatter$1,
 };
 
+const findPrettier = prettierInstallation;
 const {
-	showError: showError$2,
-	getConfigWithWorkspaceOverride: getConfigWithWorkspaceOverride$2,
-	observeConfigWithWorkspaceOverride: observeConfigWithWorkspaceOverride$1,
-	log: log$3,
+	showError,
+	getConfigWithWorkspaceOverride,
+	observeConfigWithWorkspaceOverride,
+	log,
 } = helpers;
-const { Formatter: Formatter$1 } = formatter;
+const { Formatter } = formatter;
 
 class PrettierExtension {
 	constructor() {
 		this.didAddTextEditor = this.didAddTextEditor.bind(this);
 		this.toggleFormatOnSave = this.toggleFormatOnSave.bind(this);
 		this.modulePathDidChange = this.modulePathDidChange.bind(this);
+		this.prettierConfigFileDidChange =
+			this.prettierConfigFileDidChange.bind(this);
 		this.editorWillSave = this.editorWillSave.bind(this);
 		this.didInvokeFormatCommand = this.didInvokeFormatCommand.bind(this);
-		this.didInvokeFormatSelectionCommand = this.didInvokeFormatSelectionCommand.bind(
-			this
-		);
-		this.didInvokeSaveWithoutFormattingCommand = this.didInvokeSaveWithoutFormattingCommand.bind(
-			this
-		);
+		this.didInvokeFormatSelectionCommand =
+			this.didInvokeFormatSelectionCommand.bind(this);
+		this.didInvokeSaveWithoutFormattingCommand =
+			this.didInvokeSaveWithoutFormattingCommand.bind(this);
 
 		this.saveListeners = new Map();
 		this.ignoredEditors = new Set();
 		this.issueCollection = new IssueCollection();
 
-		this.formatter = new Formatter$1();
+		this.formatter = new Formatter();
 	}
 
 	setupConfiguration() {
 		nova.config.remove('prettier.use-compatibility-mode');
 
-		observeConfigWithWorkspaceOverride$1(
+		observeConfigWithWorkspaceOverride(
 			'prettier.format-on-save',
 			this.toggleFormatOnSave
 		);
-		observeConfigWithWorkspaceOverride$1(
+		observeConfigWithWorkspaceOverride(
 			'prettier.module.path',
 			this.modulePathDidChange
 		);
@@ -1455,6 +1474,8 @@ class PrettierExtension {
 	start() {
 		this.setupConfiguration();
 
+		nova.fs.watch('**/.prettierrc*', this.prettierConfigFileDidChange);
+		nova.fs.watch('**/package.json', this.prettierConfigFileDidChange);
 		nova.workspace.onDidAddTextEditor(this.didAddTextEditor);
 		nova.commands.register('prettier.format', this.didInvokeFormatCommand);
 		nova.commands.register(
@@ -1469,10 +1490,10 @@ class PrettierExtension {
 
 	async startFormatter() {
 		const path =
-			getConfigWithWorkspaceOverride$2('prettier.module.path') ||
-			(await prettierInstallation());
+			getConfigWithWorkspaceOverride('prettier.module.path') ||
+			(await findPrettier());
 
-		log$3.info(`Loading prettier at ${path}`);
+		log.info(`Loading prettier at ${path}`);
 		await this.formatter
 			.start(path)
 			.catch(() =>
@@ -1483,7 +1504,7 @@ class PrettierExtension {
 	}
 
 	toggleFormatOnSave() {
-		this.enabled = getConfigWithWorkspaceOverride$2('prettier.format-on-save');
+		this.enabled = getConfigWithWorkspaceOverride('prettier.format-on-save');
 
 		if (this.enabled) {
 			nova.workspace.textEditors.forEach(this.didAddTextEditor);
@@ -1493,13 +1514,18 @@ class PrettierExtension {
 		}
 	}
 
+	async prettierConfigFileDidChange() {
+		await this.formatter.stop();
+		await this.formatter.start();
+	}
+
 	async modulePathDidChange() {
 		try {
-			this.formatter.stop();
+			await this.formatter.stop();
 			await this.startFormatter();
 		} catch (err) {
 			if (err.status === 127) {
-				return showError$2(
+				return showError(
 					'prettier-resolution-error',
 					`Can't find npm and Prettier`,
 					`Prettier couldn't be found because npm isn't available. Please make sure you have Node installed. If you've only installed Node through NVM, you'll need to change your shell configuration to work with Nova. See https://library.panic.com/nova/environment-variables/`
@@ -1508,7 +1534,7 @@ class PrettierExtension {
 
 			console.error('Unable to start prettier service', err, err.stack);
 
-			return showError$2(
+			return showError(
 				'prettier-resolution-error',
 				`Unable to start Prettier`,
 				`Please check the extension console for additional logs.`
@@ -1555,7 +1581,7 @@ class PrettierExtension {
 			this.issueCollection.set(editor.document.uri, issues);
 		} catch (err) {
 			console.error(err, err.stack);
-			showError$2(
+			showError(
 				'prettier-format-error',
 				`Error while formatting`,
 				`"${err.message}" occurred while formatting ${editor.document.path}. See the extension console for more info.`
@@ -1564,14 +1590,14 @@ class PrettierExtension {
 	}
 }
 
-var activate = async function () {
+var activate = main.activate = async function () {
 	try {
 		const extension = new PrettierExtension();
 		extension.start();
 	} catch (err) {
 		console.error('Unable to set up prettier service', err, err.stack);
 
-		return showError$2(
+		return showError(
 			'prettier-resolution-error',
 			`Unable to start Prettier`,
 			`Please check the extension console for additional logs.`
@@ -1579,15 +1605,10 @@ var activate = async function () {
 	}
 };
 
-var deactivate = function () {
+var deactivate = main.deactivate = function () {
 	// Clean up state before the extension is deactivated
-};
-
-var main = {
-	activate: activate,
-	deactivate: deactivate
 };
 
 exports.activate = activate;
 exports.deactivate = deactivate;
-exports.default = main;
+exports["default"] = main;
